Remove published: false from template for batch new
Now that this property isn't required anymore, we can get rid of that part. Honors the feature flag for optional publish values.
